(FIX) Doubles de los modelos

+++ src/main/java/org/springframework/samples/petclinic/model/Ingrediente.java
@@ -19,7 +19,7 @@ import lombok.Setter;
 public class Ingrediente extends BaseEntity{
 	
 	@Column(name="cantidad")
-	private Integer cantidadUsualPP;
+	private Double cantidadUsualPP;
 	
 	@ManyToOne(optional=false)
 	@JoinColumn(name = "producto_id")
