+++ src/main/java/org/springframework/samples/petclinic/model/IngredientePedido.java
@@ -18,7 +18,7 @@ public class IngredientePedido extends BaseEntity{
 	
 	
 	@Column(name = "cant_pedida")
-	private Integer cantidadPedida;
+	private Double cantidadPedida;
 	
 	@ManyToOne(optional=false)
 	@JoinColumn(name = "ingrediente_id")
